Implement a collector that should allow a safe api to be exposed
The new collector works with shadow data except when doing its final mark phase, and only alters the live/dead status during that phase.

Post-mark, it only runs drop for nodes that have remained dead and it checks to ensure each node is still dead before de-allocating it in case some drop code added extra refs.

When it detects such a node it marks as a zombie. The rest of the implementation does not override this zombie status, and we attempt to de-allocate such nodes when their owning structures drop to zero refs. Because zombies can't have children (their inner data was already dropped) we don't need to trace them for cycles.