Simplify & reduce usage of unsafe in collector
Status::Zombie is not necessary, Dead + refcount is sufficient to communicate all of the same information.
We _never_ want to actually see a refcount of zero, since the holder of the last reference performs de-allocation. Use a nonzero usize and checked math to enforce this.
We also don't need to do a second trace of dead nodes, we can just trace live ones and filter dead nodes out of the candidate list if they are children. This simplifies the dead marking phase to a simple for node in nodes { mark_dead }.