[Soundness] Fix soundness hole in Gc
In retrospect it seems obvious, but by not checking outstanding borrows of data, it's possible to drop data out from under an active borrow in the presence of a trace bug.
Now borrow status is checked during collection.
As a benefit, this should allow diagnosing a limited number of trace bugs automatically.
Probably needs some more cleanup of stuff later, but at least the bug is fixed.

Fixes: #9