Add minimal fuzzer
It currently doesn't report an arbitrary number of times (only once or twice), but this makes it easier to track behavior over time.
It also doesn't under-report, as that's a client-controlled leak and we don't really care about that, only UB in the case of over-reporting, or leaks in the case of normal behavior.