Remove almost all unsafe
The level of unsafe and testing required in the dance between a custom RC and the collector was really bothering me.

Now there's a single unsafe call, inside of Inner::drop_data, which is very carefully guarded and should be trivial to verify.

Gc now uses an Rc internally to manage its pointers, wrapping metadata + a ManuallyDrop field to allow it to break cycles when needed.
Fuzzing has been altered to no longer over-report, as the new style of handling ManuallyDrop is not a serious risk of UB and we have cases where overreporting causes leaks. We're more concerned about leaks in correct implementations of trace now.

Validated with miri on exisiting tests + fuzzing.

Performance (such as is possible to measure on my machine) is at least not horrifically regressed, and looks like it's roughly the same.